Avoid keeping two lists of malware domains
We will sometimes have to add more domains to this list,
so it is easier to keep it only in one place. For now isGreyedOut
just calls isMalware, but we might add more reasons why a URL
should be greyed out in the future.
